+++ lib/ts/recipe/session/sessionFunctions.ts
@@ -16,9 +16,10 @@ import { getInfoFromAccessToken, sanitizeNumberInput } from "./accessToken";
 import { getPayloadWithoutVerifiying } from "./jwt";
 import STError from "./error";
 import { PROCESS_STATE, ProcessState } from "../../processState";
-import { CreateOrRefreshAPIResponse } from "./types";
+import { CreateOrRefreshAPIResponse, SessionInformation } from "./types";
 import NormalisedURLPath from "../../normalisedURLPath";
 import RecipeImplementation from "./recipeImplementation";
+import { maxVersion } from "../../utils";
 
 /**
  * @description call this to "login" a user.
@@ -237,6 +238,34 @@ export async function getSession(
 }
 
 /**
+ * @description Retrieves session information from storage for a given session handle
+ * @returns session data stored in the database, including userData and JWT payload
+ */
+export async function getSessionInformation(
+    recipeImplementation: RecipeImplementation,
+    sessionHandle: string
+): Promise<SessionInformation> {
+    let apiVersion = await recipeImplementation.querier.getAPIVersion();
+
+    if (maxVersion(apiVersion, "2.7") === "2.7") {
+        throw new Error("Please use core version >= 3.5 to call this function.");
+    }
+
+    let response = await recipeImplementation.querier.sendGetRequest(new NormalisedURLPath("/recipe/session"), {
+        sessionHandle,
+    });
+
+    if (response.status === "OK") {
+        return response;
+    } else {
+        throw new STError({
+            message: response.message,
+            type: STError.UNAUTHORISED,
+        });
+    }
+}
+
+/**
  * @description generates new access and refresh tokens for a given refresh token. Called when client's access token has expired.
  * @sideEffects calls onTokenTheftDetection if token theft is detected.
  */
@@ -350,13 +379,22 @@ export async function revokeMultipleSessions(
 }
 
 /**
+ * @deprecated use getSessionInformation() instead
  * @description: this function reads from the database every time. It provides no locking mechanism in case other processes are updating session data for this session as well, so please take of that by yourself.
  * @returns session data as provided by the user earlier
  */
 export async function getSessionData(recipeImplementation: RecipeImplementation, sessionHandle: string): Promise<any> {
+    let apiVersion = await recipeImplementation.querier.getAPIVersion();
+
+    // Call new method for >= 2.8
+    if (maxVersion(apiVersion, "2.7") !== "2.7") {
+        return (await getSessionInformation(recipeImplementation, sessionHandle)).userDataInDatabase;
+    }
+
     let response = await recipeImplementation.querier.sendGetRequest(new NormalisedURLPath("/recipe/session/data"), {
         sessionHandle,
     });
+
     if (response.status === "OK") {
         return response.userDataInDatabase;
     } else {
@@ -389,6 +427,7 @@ export async function updateSessionData(
 }
 
 /**
+ * @deprecated use getSessionInformation() instead
  * @returns jwt payload as provided by the user earlier
  */
 export async function getJWTPayload(recipeImplementation: RecipeImplementation, sessionHandle: string): Promise<any> {