+++ lib/ts/supertokens.ts
@@ -24,6 +24,7 @@ import {
     assertThatBodyParserHasBeenUsed,
     validateTheStructureOfUserInput,
     removeServerlessCache,
+    maxVersion,
 } from "./utils";
 import { Querier } from "./querier";
 import RecipeModule from "./recipeModule";
@@ -240,5 +241,55 @@ export default class SuperTokens {
             });
         });
         return Array.from(headerSet);
+    };
+
+    getUserCount = async (includeRecipeIds?: string[]): Promise<number> => {
+        let querier = Querier.getNewInstanceOrThrowError(this.isInServerlessEnv, undefined);
+        let apiVersion = await querier.getAPIVersion();
+        if (maxVersion(apiVersion, "2.7") === "2.7") {
+            throw new Error(
+                "Please use core version >= 3.5 to call this function. Otherwise, you can call <YourRecipe>.getUserCount() instead (for example, EmailPassword.getUserCount())"
+            );
+        }
+        let includeRecipeIdsStr = undefined;
+        if (includeRecipeIds !== undefined) {
+            includeRecipeIdsStr = includeRecipeIds.join(",");
+        }
+        let response = await querier.sendGetRequest(new NormalisedURLPath("/users/count"), {
+            includeRecipeIds: includeRecipeIdsStr,
+        });
+        return Number(response.count);
+    };
+
+    getUsers = async (input: {
+        timeJoinedOrder: "ASC" | "DESC";
+        limit?: number;
+        paginationToken?: string;
+        includeRecipeIds?: string[];
+    }): Promise<{
+        users: { recipeId: string; user: any }[];
+        nextPaginationToken?: string;
+    }> => {
+        let querier = Querier.getNewInstanceOrThrowError(this.isInServerlessEnv, undefined);
+        let apiVersion = await querier.getAPIVersion();
+        if (maxVersion(apiVersion, "2.7") === "2.7") {
+            throw new Error(
+                "Please use core version >= 3.5 to call this function. Otherwise, you can call <YourRecipe>.getUsersOldestFirst() or <YourRecipe>.getUsersNewestFirst() instead (for example, EmailPassword.getUsersOldestFirst())"
+            );
+        }
+        let includeRecipeIdsStr = undefined;
+        if (input.includeRecipeIds !== undefined) {
+            includeRecipeIdsStr = input.includeRecipeIds.join(",");
+        }
+        let response = await querier.sendGetRequest(new NormalisedURLPath("/users"), {
+            includeRecipeIds: includeRecipeIdsStr,
+            timeJoinedOrder: input.timeJoinedOrder,
+            limit: input.limit,
+            paginationToken: input.paginationToken,
+        });
+        return {
+            users: response.users,
+            nextPaginationToken: response.nextPaginationToken,
+        };
     };
 }