+++ test/with-typescript/index.ts
@@ -37,6 +37,8 @@ Supertokens.init({
                                 revokeSession: session.revokeSession,
                                 updateJWTPayload: session.updateJWTPayload,
                                 updateSessionData: session.updateSessionData,
+                                getExpiry: session.getExpiry,
+                                getTimeCreated: session.getTimeCreated,
                             };
                         },
                         getAllSessionHandlesForUser: originalImpl.getAllSessionHandlesForUser,
@@ -50,6 +52,7 @@ Supertokens.init({
                         updateSessionData: originalImpl.updateSessionData,
                         getAccessTokenLifeTimeMS: originalImpl.getAccessTokenLifeTimeMS,
                         getRefreshTokenLifeTimeMS: originalImpl.getRefreshTokenLifeTimeMS,
+                        getSessionInformation: originalImpl.getSessionInformation,
                     };
                 },
             },
@@ -263,6 +266,63 @@ ThirdPartyEmailPassword.init({
                         }
                     }
                     return response;
+                },
+            };
+        },
+    },
+});
+
+async function f() {
+    let n: number = await Supertokens.getUserCount(["a", "b"]);
+    let n2: number = await Supertokens.getUserCount();
+
+    await Supertokens.getUsersOldestFirst({
+        includeRecipeIds: [""],
+        limit: 1,
+        paginationToken: "",
+    });
+
+    await Supertokens.getUsersNewestFirst({
+        includeRecipeIds: [""],
+        limit: 1,
+        paginationToken: "",
+    });
+}
+
+EmailPassword.init({
+    override: {
+        apis: (originalImplementation) => {
+            return {
+                ...originalImplementation,
+                signInPOST: async (input) => {
+                    let formFields = input.formFields;
+                    let options = input.options;
+                    let email = formFields.filter((f) => f.id === "email")[0].value;
+                    let password = formFields.filter((f) => f.id === "password")[0].value;
+
+                    let response = await options.recipeImplementation.signIn({ email, password });
+                    if (response.status === "WRONG_CREDENTIALS_ERROR") {
+                        return response;
+                    }
+                    let user = response.user;
+
+                    let origin = options.req.headers["origin"];
+
+                    let isAllowed = false; // TODO: check if this user is allowed to sign in via their origin..
+
+                    if (isAllowed) {
+                        // import Session from "supertokens-node/recipe/session"
+                        await Session.createNewSession(options.res, user.id);
+                        return {
+                            status: "OK",
+                            user,
+                        };
+                    } else {
+                        // on the frontend, this will display incorrect email / password combination
+                        return {
+                            status: "WRONG_CREDENTIALS_ERROR",
+                        };
+                    }
                 },
             };
         },